extract ManageSubscriptionsButtonsView
