+++ RevenueCatUI/CustomerCenter/Views/ManageSubscriptionsButtonsView.swift
@@ -1,0 +1,77 @@
+//
+//  Copyright RevenueCat Inc. All Rights Reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
+//  ManageSubscriptionsButtonsView.swift
+//
+//  Created by Cesar de la Vega on 2/12/24.
+
+import Foundation
+import RevenueCat
+import SwiftUI
+
+#if os(iOS)
+
+@available(iOS 15.0, *)
+@available(macOS, unavailable)
+@available(tvOS, unavailable)
+@available(watchOS, unavailable)
+struct ManageSubscriptionsButtonsView: View {
+
+    @ObservedObject
+    var viewModel: ManageSubscriptionsViewModel
+    @Binding
+    var loadingPath: CustomerCenterConfigData.HelpPath?
+    @Environment(\.openURL)
+    var openURL
+
+    @Environment(\.localization)
+    private var localization: CustomerCenterConfigData.Localization
+
+    var body: some View {
+        let filteredPaths = self.viewModel.screen.paths.filter { path in
+#if targetEnvironment(macCatalyst)
+            return path.type == .refundRequest
+#else
+            return path.type != .unknown
+#endif
+        }
+        ForEach(filteredPaths, id: \.id) { path in
+            ManageSubscriptionButton(path: path, viewModel: self.viewModel)
+        }
+    }
+
+}
+
+@available(iOS 15.0, *)
+@available(macOS, unavailable)
+@available(tvOS, unavailable)
+@available(watchOS, unavailable)
+struct ManageSubscriptionButton: View {
+
+    let path: CustomerCenterConfigData.HelpPath
+    @ObservedObject var viewModel: ManageSubscriptionsViewModel
+
+    @Environment(\.appearance) private var appearance: CustomerCenterConfigData.Appearance
+
+    var body: some View {
+        AsyncButton(action: {
+            await self.viewModel.determineFlow(for: path)
+        }, label: {
+            if self.viewModel.loadingPath?.id == path.id {
+                TintedProgressView()
+            } else {
+                Text(path.title)
+            }
+        })
+        .disabled(self.viewModel.loadingPath != nil)
+    }
+
+}
+
+#endif
